refactor: type renames
