+++ src/stores/spaces.ts
@@ -2,9 +2,8 @@ import { defineStore, skipHydrate } from "pinia";
 import { spaceFilter } from "~/lib/filter-spaces";
 import type { BuildingI18n } from "~/types/Building";
 import type { Filters } from "~/types/filters";
-import type { Space, SpaceI18n } from "~/types/Space";
+import type { Space, SpaceI18n, Room, RoomI18n } from "~/types/Space";
 import calculateOccupancy from "../lib/calculate-occupancy";
-import { Room, RoomI18n } from "~/types/Room";
 import { useMapStore } from "./map";
 
 export type Selection =
