+++ src/types/Space.ts
@@ -2,7 +2,40 @@ import type { SpaceFeatures } from "./Filters";
 import type { OpeningHours } from "./OpeningHours";
 import type { Building } from "./Building";
 
-export interface Space extends SpaceI18n {
+export type Space = Room | RoomZone;
+
+export interface Room extends RoomI18n {
+  name: string;
+  slug: string;
+  building: Building;
+}
+
+export interface RoomI18n {
+  floor: string;
+  seats: number;
+  tables: number;
+  spaces: number;
+  latitude: number;
+  longitude: number;
+  roomId: string;
+  realEstateNumber: string;
+  facilities: SpaceFeatures;
+  buildingNumber: number;
+  i18n: Record<string, { name: string }>;
+  openingHours: OpeningHours[];
+  imageUrl?: string | undefined; //TODO: not there
+  activeDevices?: number;
+}
+
+export type CsvRoomData = Omit<
+  RoomI18n,
+  "openingHours" | "imageUrl" | "activeDevices"
+> & {
+  exchangeBuildingId: string;
+  exchangeRoomId: string;
+};
+
+export interface RoomZone extends SpaceI18n {
   name: string;
   slug: string;
   building: Building;
